move validations to their own folder
