+++ validations/index.ts
@@ -1,5 +1,5 @@
 // ANCHOR: Export User Login Validation
-export { userLoginValidation } from "./modules/userLoginValidation";
+export { userLoginValidation } from "./modules/user/userLoginValidation";
 
 // ANCHOR: Export User Register Validation
-export { userRegisterValidation } from "./modules/userRegisterValidation";
+export { userRegisterValidation } from "./modules/user/userRegisterValidation";
